+++ lib/src/models/pass_file.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:pass_flutter/src/models/pass_image.dart';
 import 'package:pass_flutter/src/models/pass_json/pass_json.dart';
+import 'package:pass_flutter/src/pass_core.dart';
 
 /// Parsed Apple Wallet pass file
 class PassFile {
@@ -39,9 +40,15 @@ class PassFile {
   /// A JSON dictionary that defines the pass.
   PassJson pass;
 
+  /// Creates a new [PassFile]
   PassFile(String id, File passFile, Directory passDirectory) {
     this.id = id;
     this.file = passFile;
     this.directory = passDirectory;
+  }
+
+  /// Delete current pass file from memory
+  void delete() {
+    PassIo().delete(this.directory, this.file);
   }
 }