+++ lib/src/pass_core.dart
@@ -6,9 +6,7 @@ import 'package:archive/archive.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-import 'package:pass_flutter/src/models/pass_image.dart';
-import 'package:pass_flutter/src/models/pass_file.dart';
-import 'package:pass_flutter/src/models/pass_json/pass_json.dart';
+import 'package:pass_flutter/pass_flutter.dart';
 import 'package:path/path.dart' as path;
 import 'package:path_provider/path_provider.dart';
 import 'package:uuid/uuid.dart';
@@ -24,42 +22,33 @@ class Pass {
   static const String _channelName = 'pass_flutter';
   static const MethodChannel _channel = MethodChannel(_channelName);
 
-  /// Save pass file from [url] to internal memory,
-  /// parse and return [PassFile]
-  Future<PassFile> getFromUrl(String url) async {
-    PassFile passFile = await _PassIo().createOrGetPass();
-    String pathToSave = passFile.file.path;
-    Response<ResponseBody> responce = await Dio().download(url, pathToSave);
-    if (responce.statusCode == 200) {
-      return await _PassParser().parse(passFile);
-    }
-    throw ('Unable to download pass file at specified url');
-  }
-
   /// Return all saved pass files from internal memory in List of [PassFile]
   Future<List<PassFile>> getAllSaved() async {
-    List<PassFile> parsedPasses = [];
-    Directory passesDir = await _PassIo().getPassesDir();
-    List<FileSystemEntity> passes = await passesDir.list().toList();
-    for (var entity in passes) {
-      if (entity is File) {
-        String passId = path.basenameWithoutExtension(entity.path);
-        PassFile passFile = await _PassIo().createOrGetPass(passId: passId);
-        try {
-          passFile = await _PassParser().parse(passFile);
-          parsedPasses.add(passFile);
-        } catch (e) {
-          debugPrint('Error parse pass file - ${passFile.file.path}');
-          this.delete(passFile);
-        }
-      }
-    }
+    List<PassFile> parsedPasses = await PassIo().getAllSaved();
     return parsedPasses;
   }
 
-  /// Delete all files and folders for [passFile] from internal memory
-  delete(PassFile passFile) {
-    _PassIo().delete(passFile.directory, passFile.file);
+  /// Save pass file from [urlToPass] to internal memory, parse and return [PassFile]
+  Future<PassFile> saveFromUrl({@required String urlToPass}) async {
+    PassFile pass = await PassIo().saveFromUrl(url: urlToPass);
+    return pass;
+  }
+
+  /// Fetch preview of pass file from [urlToPass], parse and return [PassFile]
+  Future<PreviewPassFile> fetchPreviewFromUrl({
+    @required String urlToPass,
+  }) async {
+    PreviewPassFile pass = await PassIo().fetchPreviewFromUrl(
+      url: urlToPass,
+    );
+    return pass;
+  }
+
+  /// Delete all files and folders for [passFile] from internal memory and return saved passes
+  Future<List<PassFile>> delete(PassFile passFile) async {
+    PassIo().delete(passFile.directory, passFile.file);
+    List<PassFile> parsedPasses = await PassIo().getAllSaved();
+    return parsedPasses;
   }
 
   /// Platform version