+++ CHANGELOG.md
@@ -1,13 +1,23 @@
+## 1.0.3 - 03.04.2020
+ - **BREAKING:** change method name getFromUrl -> saveFromUrl and url now passes in `urlToPass`
+ - add method `fetchPreviewFromUrl` for get preview from url
+ - add method `PreviewPassFile.save` for save preview to app internal memory and deleting from temp directory
+ - add documentation for all public api members
+ - add `delete` method for deleting current pass file to class PassFile and PreviewPassFile
+ - update return in `Pass.delete` method, after deleting now return list of saved pass files
+ - enable stricter type checks
+ - refactoring and optimization
+
 ## 1.0.2 - 28.03.2020
- - fix bug with maxDistance and relevantDate in pass json
- - fix bug with auxiliaryFields, primaryFields, transitType
+ - fix bug with `maxDistance` and `relevantDate` in pass json
+ - fix bug with `auxiliaryFields`, `primaryFields`, `transitType`
  - add equatable for comparisons pass.json data
  - add documentation for Fields
 
 ## 1.0.1 - 13.03.2020
  - fix code format
- - add altitude to Location Dictionary Keys
- - add maxDistance to Relevance Keys
+ - add `altitude` to Location Dictionary Keys
+ - add `maxDistance` to Relevance Keys
  - add documentation to Location Dictionary Keys model
  - add documentation to Barcode Dictionary Keys model
  - add documentation to Pass Structure Dictionary model